Adding first implementation of bitcoin backend generalizzation
