+++ lnprototest/backend/__init__.py
@@ -1,0 +1,7 @@
+from .backend import Backend
+from .bitcoind import Bitcoind
+
+__all__ = [
+    "Backend",
+    "Bitcoind"
+]
